fix(ui): show container names and location
- on asset scans page
- on findings page
