+++ ui/src/layout/Assets/AssetsTable.js
@@ -8,7 +8,7 @@ import { APIS } from 'utils/systemConsts';
 import { getFindingsColumnsConfigList, getVulnerabilitiesColumnConfigItem, getAssetColumnsFiltersConfig,
     findingsColumnsFiltersConfig, vulnerabilitiesCountersColumnsFiltersConfig, formatTagsToStringsList, formatDate} from 'utils/utils';
 import { useFilterDispatch, useFilterState, setFilters, FILTER_TYPES } from 'context/FiltersProvider';
-import { getAssetName } from './utils';
+import { getAssetName } from 'utils/utils';
 
 const TABLE_TITLE = "assets";
