--- ui/src/layout/Assets/utils.js
@@ -1,16 +1,0 @@
-export function getAssetName(assetInfo) {
-    switch (assetInfo.objectType) {
-        case "VMInfo":
-            return assetInfo.instanceID;
-        case "PodInfo":
-            return assetInfo.podName;
-        case "DirInfo":
-            return assetInfo.dirName;
-        case "ContainerImageInfo":
-            return assetInfo.imageID;
-        case "ContainerInfo":
-            return assetInfo.containerName;
-        default:
-            return assetInfo.id;
-    }
-}
