+++ ui/src/utils/utils.js
@@ -184,3 +184,20 @@ export const getAssetColumnsFiltersConfig = (props) => {
 }
 
 export const formatTagsToStringsList = tags => tags?.map(({key, value}) => `${key}=${value}`);
+
+export function getAssetName(assetInfo) {
+    switch (assetInfo.objectType) {
+        case "VMInfo":
+            return assetInfo.instanceID;
+        case "PodInfo":
+            return assetInfo.podName;
+        case "DirInfo":
+            return assetInfo.dirName;
+        case "ContainerImageInfo":
+            return assetInfo.imageID;
+        case "ContainerInfo":
+            return assetInfo.containerName;
+        default:
+            return assetInfo.id;
+    }
+}
